+++ pgwatch3_rpc_server/src/go.mod
@@ -1,3 +1,3 @@
-module main
+module pgwatch3_rpc_receiver
 
 go 1.22.1
