+++ pgwatch3_rpc_server/src/main.go
@@ -1,54 +1,13 @@
 package main
 
 import (
-	"errors"
 	"flag"
 	"log"
 	"net"
 	"net/http"
 	"net/rpc"
+    . "pgwatch3_rpc_receiver/sinks"
 )
-
-func (receiver *Receiver) UpdateMeasurements(msg *MeasurementMessage, logMsg *string) error {
-	log.Print("Received metrics: " + msg.DBName)
-	if len(msg.DBName) == 0 {
-		log.Print("Empty")
-		return nil
-	}
-	if receiver.sink_type == CSV {
-		writer := new(CSVReceiver)
-		err := writer.UpdateMeasurements(msg, logMsg, receiver.storage_folder, receiver)
-		if err != nil {
-			return err
-		}
-	} else if receiver.sink_type == TEXT {
-		writer := new(TextReceiver)
-		err := writer.UpdateMeasurements(msg, logMsg)
-		if err != nil {
-			return err
-		}
-	} else {
-		return errors.New("No writer was specified")
-	}
-	return nil
-}
-
-// Gets the sync request from pgwatch3 and adds it to the receiver.SyncChannel
-func (receiver *Receiver) SyncMetricSignal(syncReq *SyncReq, logMsg *string) error {
-	log.Print("RECEVIED SIGNAL")
-	go receiver.PopulateChannel(syncReq)
-	log.Print("Logged signal")
-	return nil
-}
-
-func (receiver *Receiver) PopulateChannel(syncReq *SyncReq) {
-	receiver.SyncChannel <- *syncReq
-}
-
-func (receiver *Receiver) GetSyncChannelContent() SyncReq {
-	content := <-receiver.SyncChannel
-	return content
-}
 
 func main() {
 
@@ -68,14 +27,14 @@ func main() {
 
 	server.SyncChannel = make(chan SyncReq, 10)
 	if *receiverType == "csv" {
-		server.sink_type = CSV
-		server.storage_folder = *storage_folder
+		server.SinkType = CSV
+		server.StorageFolder = *storage_folder
 	} else if *receiverType == "text" {
 		// Only for testing
-		server.sink_type = TEXT
+		server.SinkType = TEXT
 	} else {
 		// Throw Error
-		server.sink_type = NONE
+		server.SinkType = NONE
 		log.Fatal("[ERROR]: No Sink Type was provided. Please use the --type option")
 		return
 	}
