+++ pgwatch3_rpc_server/src/sinks/csv_receiver.go
@@ -1,4 +1,4 @@
-package main
+package sinks
 
 import (
 	"encoding/csv"
